requirements.txt
numpy==1.19.2
Pillow==8.4.0
streamlit==1.1.0
tensorflow-cpu
